Move cpa and dpa to distinguishers module
