+++ src/distinguishers/mod.rs
@@ -1,0 +1,3 @@
+pub mod cpa;
+pub mod cpa_normal;
+pub mod dpa;
